+++ packages/editor-pro/js/index.js
@@ -1,3 +1,4 @@
 // @flow
 
 export * from './extensions';
+export * from './canvas-editor';