+++ packages/blockera-pro-admin/CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Unreleased
 
+### New Features
+- Added support for customizing your website's design across all screen sizes (mobile, tablet, desktop, and any other custom breakpoints).
+- Enabled access to responsive design controls after activating the Pro version license.
+- Improved the responsive design system to allow easier management of breakpoint settings.
+
 ## 1.1.0 (2025-06-12)
 
 ### New Features